Move structs for json unmarshaling to own file
